added bootstrap css and simplified layout

+++ php/grubhub/index.php
@@ -1,6 +1,7 @@
 <?php
 
-define('URL', 'http://www.kevinx.net/labs/php/grubhub');
+//define('URL', 'http://www.kevinx.net/labs/php/grubhub');
+define('URL', 'http://kevinx.local/labs/php/grubhub');
 
 require 'simple_html_dom.php';
 
@@ -85,6 +86,7 @@ class GrubHubApp
             curl_setopt($ch, CURLOPT_COOKIEJAR, $this->cookieJarFilename());
             $res = curl_exec($ch);
             curl_close($ch);
+            
             if($res !== false && strpos($res, 'Internal Server Error') === false)
             {
                 $_SESSION['user_hash'] = md5($_POST['username']);
@@ -109,16 +111,14 @@ class GrubHubApp
     {
         $this->loginRequired();
 
-        $top5 = array();
         $pastOrders = array();
-        $rankedOrders = array();
-        $recentOrders = array();
+        $aggOrders = array();
 
-        // all past orders
-        $sql = "select name, total, date from past_order where user_hash='%s' order by date desc";
-        $sql = sprintf($sql, $_SESSION['user_hash']);
-        $query = $this->db->query($sql);
-        foreach($query->fetchAll() as $row)
+        // past orders
+        $sql = "select name, total, date from past_order where user_hash=? order by date desc";
+        $stmt = $this->db->prepare($sql);
+        $stmt->execute(array($_SESSION['user_hash']));
+        foreach($stmt->fetchAll() as $row)
         {
             $order = new PastOrder();
             $order->name = $row['name'];
@@ -127,56 +127,53 @@ class GrubHubApp
             $pastOrders[] = $order;
         }
 
-        // ranked
-        $sql = "
-            select name, count(*) as timesOrdered, sum(total) as total
-            from past_order
-            where user_hash='%s'
-            group by name
-            order by timesOrdered desc";
-        $sql = sprintf($sql, $_SESSION['user_hash']);
-        $query = $this->db->query($sql);
-        $n = 0;
-        foreach($query->fetchAll() as $row)
+        // aggregate
+        $sql = "select name, count(*) as timesOrdered, sum(total) as total, max(date) as lastDate
+                from past_order
+                where user_hash=?
+                group by name
+                order by timesOrdered desc";
+        $stmt = $this->db->prepare($sql);
+        $stmt->execute(array($_SESSION['user_hash']));
+        $i = 0;
+        foreach($stmt->fetchAll() as $row)
         {
-            $order = new RankedOrder();
+            $order = new AggregateOrder();
             $order->name = $row['name'];
             $order->timesOrdered = $row['timesOrdered'];
             $order->total = $row['total'];
-            if($n < 5)
-            {
-                $top5[] = $order->name;
-            }
-            $rankedOrders[] = $order;
-            $n++;
-        }
-
-        // recent orders
-        $sql = "
-            select name, max(date) as lastDate
-            from past_order
-            where user_hash='%s'
-            group by name
-            order by lastDate desc";
-        $sql = sprintf($sql, $_SESSION['user_hash']);
-        $query = $this->db->query($sql);
-        foreach($query->fetchAll() as $row)
-        {
-            $order = new PastOrder();
-            $order->name = $row['name'];
-            $order->date = strtotime($row['lastDate']);
-            if(in_array($order->name, $top5))
+            $order->lastDate = strtotime($row['lastDate']);
+            if($i < 5)
             {
                 $order->name = '<strong>' . $order->name . '</strong>';
             }
-            $recentOrders[] = $order;
+            $aggOrders[] = $order;
+            $i++;
         }
 
+        usort($aggOrders, array($this, '_sortByDate'));
+
+        $grandTotal = array_sum(array_map(array($this, '_orderTotals'), $aggOrders));
+        
         include 'templates/dashboard.php';
     }
-
+    
+    function _sortByDate($a, $b)
+    {
+        return $a->lastDate == $b->lastDate ? 0 : $a->lastDate > $b->lastDate ? -1 : 1;
+    }
+    
+    function _orderTotals($order)
+    {
+        return $order->total;
+    }
+    
     function showImport()
     {
+        if(!isset($_POST['import']))
+        {
+            return;
+        }
         $this->loginRequired();
 
         set_time_limit(0);
@@ -208,7 +205,7 @@ class GrubHubApp
             $html = str_get_html($res);
             $htmlOrders = $html->find('.past_orders');
 
-            $n = 0;
+            $i = 0;
             $this->db->beginTransaction();
             foreach($htmlOrders as $htmlOrder)
             {
@@ -216,9 +213,9 @@ class GrubHubApp
                 $name = trim($name);
 
                 $date = $htmlOrder->find('.yourGrubHubColumnWrapper', 0)->plaintext;
+				$date = preg_replace('/[a-z]/i', '', $date);
                 $date = trim($date);
-                list($month, $day, $year, $hour, $minute, $ampm) = sscanf($date, 'Ordered On %s %d %d at %d:%d %s');
-                $time = strtotime("$month $day $year $hour:$minute $ampm");
+                $time = strtotime($date);
                 $date = date('Y-m-d H:i', $time);
 
                 $total = $htmlOrder->find('.totalRow .amountColumn', 0)->plaintext;
@@ -241,10 +238,10 @@ class GrubHubApp
                 echo "Imported {$order->name}...\n";
                 flush();
 
-                $n++;
+                $i++;
             }
             $this->db->commit();
-            if($n < 24)
+            if($i < 24)
             {
                 echo "Nothing else to import!\n";
                 flush();
@@ -252,18 +249,25 @@ class GrubHubApp
             }
         }
 
-        echo 'Complete!</pre>';
+        echo 'Complete!</pre><hr />';
         echo '<p><a href="dashboard">Return to Dashboard</a></p>';
     }
 
     function clearData()
     {
+        if(!isset($_POST['clear']))
+        {
+            return;
+        }
         $this->loginRequired();
         $sql = "delete from past_order where user_hash='%s'";
         $sql = sprintf($sql, $_SESSION['user_hash']);
         $this->db->query($sql);
-        header('Location: ' . URL . '/dashboard');
-        exit;
+        if(file_exists($this->cookieJarFilename()))
+        {
+            unlink($this->cookieJarFilename());
+        }
+        $this->doLogout();
     }
 
     function showDisclaimer()
@@ -292,50 +296,38 @@ class PastOrder
     {
         return date('M j Y g:i A', $this->date);
     }
-
-    function daysAgo()
-    {
-        return round((time() - $this->date) / 60 / 60 / 24);
-    }
-
+    
     function exists($db)
     {
-        $sql = "
-            select 1 from past_order po
-            where user_hash = '%s' and po.name = '%s' and po.total = '%s' and po.date = '%s'
-        ";
-        $sql = sprintf($sql,
+        $sql = "select 1 from past_order where user_hash=? and name=? and total=? and date=?";
+        $stmt = $db->prepare($sql);
+        $stmt->execute(array(
             $_SESSION['user_hash'],
-            sqlite_escape_string($this->name),
-            sqlite_escape_string($this->total),
-            sqlite_escape_string($this->date));
-        $query = $db->query($sql);
-        $res = $query->fetch();
+            $this->name,
+            $this->total,
+            $this->date));
+        $res = $stmt->fetch();
         return $res !== false && count($res) > 0;
     }
 
     function save($db)
     {
-        $sql = "
-            insert into past_order (user_hash, name, total, date)
-            values('%s', '%s', '%s', '%s')
-        ";
-        $sql = sprintf($sql,
+        $sql = "insert into past_order (user_hash, name, total, date) values(?, ?, ?, ?)";
+        $stmt = $db->prepare($sql);
+        $stmt->execute(array(
             $_SESSION['user_hash'],
-            sqlite_escape_string($this->name),
-            sqlite_escape_string($this->total),
-            sqlite_escape_string($this->date));
-        $db->query($sql);
+            $this->name,
+            $this->total,
+            $this->date));
     }
 }
 
-class RankedOrder
+class AggregateOrder
 {
     var $name;
-    // Number of visits
     var $timesOrdered;
-    // Total money spent
     var $total;
+    var $lastDate;
 
     function name()
     {
@@ -350,6 +342,11 @@ class RankedOrder
     function timesOrdered()
     {
         return $this->timesOrdered;
+    }
+    
+    function daysAgo()
+    {
+        return round((time() - $this->lastDate) / 60 / 60 / 24);
     }
 }
